Move remaining realm classes out of `Stores` namespace
