+++ osu.Game/Skinning/SkinModelManager.cs
@@ -14,7 +14,6 @@ using osu.Game.Database;
 using osu.Game.Extensions;
 using osu.Game.IO;
 using osu.Game.IO.Archives;
-using osu.Game.Stores;
 using Realms;
 
 #nullable enable
